remove duplicate functions
