+++ .gitignore
@@ -1,4 +1,5 @@
 .env
+notes
 
 # Logs
 logs
