+++ features/recognize.js
@@ -3,6 +3,8 @@ const recognition = require("../service/recognition");
 const winston = require("../winston");
 const { SlackError, GratitudeError } = require("../service/errors");
 const { reactionMatches } = require("../middleware");
+const { userInfo } = require("../service/apiwrappers");
+const { handleSlackError, handleGratitudeError, handleGenericError, sendNotificationToReceivers } = require("../service/messageutils");
 
 const { recognizeEmoji, reactionEmoji } = config;
 
@@ -110,20 +112,6 @@ async function respondToRecognitionReaction({ event, client }) {
   ]);
 }
 
-// API Wrappers
-
-async function userInfo(client, userId) {
-  const response = await client.users.info({ user: userId });
-  if (response.ok) {
-    return response.user;
-  }
-  throw new SlackError(
-    "users.info",
-    response.error,
-    `Something went wrong while sending recognition. When retreiving user information from Slack, the API responded with the following error: ${response.message} \n Recognition has not been sent.`
-  );
-}
-
 async function messageReactedTo(client, message) {
   const response = await client.conversations.replies({
     channel: message.item.channel,
@@ -138,55 +126,4 @@ async function messageReactedTo(client, message) {
     response.error,
     `Something went wrong while sending recognition. When retreiving message information from Slack, the API responded with the following error: ${response.message} \n Recognition has not been sent.`
   );
-}
-
-// Response Message Utils
-
-async function handleSlackError(client, message, error) {
-  winston.error("Slack API returned an error response", {
-    apiMethod: error.apiMethod,
-    apiError: error.apiError,
-  });
-  return client.chat.postEphemeral({
-    channel: message.channel,
-    user: message.user,
-    text: error.userMessage,
-  });
-}
-
-async function handleGratitudeError(client, message, error) {
-  winston.info("Rejected gratitude request as invalid", {
-    gratitudeErrors: error.gratitudeErrors,
-  });
-  const errorString = error.gratitudeErrors.join("\n");
-  return client.chat.postEphemeral({
-    channel: message.channel,
-    user: message.user,
-    text: `Sending gratitude failed with the following error(s):\n${errorString}`,
-  });
-}
-
-async function handleGenericError(client, message, error) {
-  winston.error("Slack API returned an error response", {
-    error,
-  });
-  const userMessage = `An unknown error occured in Gratibot: ${error.message}`;
-  return client.chat.postEphemeral({
-    channel: message.channel,
-    user: message.user,
-    text: userMessage,
-  });
-}
-
-async function sendNotificationToReceivers(client, gratitude) {
-  for (let i = 0; i < gratitude.receivers.length; i++) {
-    await client.chat.postMessage({
-      channel: gratitude.receivers[i].id,
-      text: `You earned a ${recognizeEmoji}.`,
-      ...(await recognition.receiverSlackNotification(
-        gratitude,
-        gratitude.receivers[i].id
-      )),
-    });
-  }
 }
